Added job: Military Hacker [donator]
Made this because i was bored and for this reason:

Alot of people make bases with a shitload of fading doors in one entrance. you cant enter them with a warrant because then, the prop will just get supported by the prop behind it/against it.
I think this should be fair against the criminals as well! and anyways, a new job couldnt hurt right ;)